update
- FileInfo widget code cleanup
- FileInfo widget additional actions for setting cover. Allows setting
cover to tag for the displayed item or items from its album.
- FileInfo has sort alphanumeric/semantic order for info fields
- DragUtil uses future Fut instead of Supplier. This allows us to query
data in case the future is completed from the very beginning. And in
effect, display this data in ActionPane after drag & drop. In the
future, there should be support for displaying the progress or status.
Question remains whether to run the future after user picks action or do
it right away to preload data.
- more