update
- action chooser improved. Layout fixed, added hide on action done
settings, as well as button for it in chooser header controls. Actions
are now FastActin & SlowAction. The later takes a Fut (future) and
'consumes' it, but doesnt run it on its own. It is returned back to
action chooser, which can then insert progress monitoring and wait for
it to finish.
- CheckIcon uses Property<Boolean> not BooleanProperty for selection.
Allows binding with any property as it should.