DragPane bugfixes
- dragpane consistent with drag accepting, pane displays only when the
node it is installed for accepted the drag. This requires using a
predicate consistent with the node's drag_over handler accepting
behavior. Added complexity, but the only way.
- dragpane installing supports Node (not just Pane)
- dragpane doesnt interfere with existing layout now - using
setManaged(false)
- dragpane action description text can be supplied when drag enters -
text can be build dynamically to reflect some settings
- dragpane css support
- audio drags now accepted if drag contains directory. Directory is not
searched though, so hasAudio() does not guarantee presence of audio
anymore. But this is necessary to support directories (e.g. adding audio
in directory to playlist) and the directory simply can not be traversed
for simple drag content check