function chain improvements
- links in chain nodes can be individually disabled if logically
possible. This is due to improved homogeneity check. Links in the chain
can be disabled not only if the chain is homogeneous, but also if it is
homogeneous locally. For example if the chain of functions contains
multiple String -> String functions in a row, these can be disabled,
because if the function is removed, the chain can still be composed into
a single function. This requires functions that know its generic type at
runtime...
- TypeAwareF as a simple implementation of function that knows its
generic type
- stronger function identity preservation. Added IDENTITY function
(x->x) instance.  IDENTITY function in a function chain always makes a
homogeneous link (link which can be disabled). Note that it is
impossible to even rely on TypeAware wrapper of identity function,
because its input and output would actually be Object, which is correct,
but for an input of class C, it returns output of C, which the function
generalizes to Object. This is because 'normally' there should be 1
identity function per every class c -> c, but that is a pointless
excersize, since all of these are still single function. Hence we need
to truly rely on identity (reference checks function==IDENTITY) in case
by case basis. If it was to be implemented generally, each function
would require a getOutTypeForInType(Class intype) method, where identity
would return the argument. I feel as if that would polluting the code
even more, so for now that will not be implemented.
- bugfixes