Reduce type parameters of Try Monad to one
The methods themselves should not be responsible for logging, that
should be handled as high-up as possible.
To accomplish that, Try will now contain full Error information,
while it previously only carried a Message or an Exception. This way,
the receiver of the Try instance can decide how to proceed, that
externalises the logging. Try now also has a default log method to
log its state to the corresponding levels.