Fix issues regarding ObservableListConfigField (WIP)
I srsly dont have time for this. It's cursed. It worked, but only for
'editable' types, i.e., those not requiring onChange handler to refresh
the list when item is added. Making it work for the 'addable' type is
considerably more difficult. It almost works right now, but with the
newly added support for nullable types, its even more difficult to
cover all the edge cases.