Improve Converter function chain null handling
This seemingly minor change makes the function transformation chain much more robust by enriching function supplier with not only index value, but also the very output of the previous transformation! This makes few additional tricks possible that otherwise are not (not even with fully reified type tokens), such as
- comparing outputs between transformation steps to remove unneeded Transformation.Manual step
- narrow down nullable function output type to non null when null is not in the results. This is special case of the bellow.
- narrow down function output type to any subtype by inspecting common supertype among the results, which allows then chaining functions that user would expect to be chainable (because user intuitively understands the type of the output value to be determining factor in function determination instead of function output type, which may be too broad). The function output type is still required, as it is the only means of reifying the types, however for non-generic classes and some common generic classes it is possible to further narrow down the common type from the function output type, such as Int from Number, or Int from Int?. Thus, the type tokens of the function output type and output value class examination goes hand in hand.

The actual narrowing down is yet to be implemented, but the null handling already works as it is supposed to (without providing non null functions on nullable outputs and being forced to propagate null across the rest of the chain). This fills an important blind spot in the function selection algorithm.
The use of index in the function provider natural fit and even improves the API. The use of the intermediate outputs, while not trivial to implement, is also intuitive and will makes it possible to
- avoid re-computation costs (at the cost of memory)
- use non-deterministic functions
- use transformation chain projection, where user could, through UI, ignore a tail of the transformation chain, to quickly see the intermediate results inside the chain.