Implement OverlayPane hideOnFocus
Very useful feature that simplifies UX and even code. Hiding now just works.

This was tricky to implement due to the fact that there is Node mode and Stage mode and we need to maintain whether any content is focused within the showing and focused owning Node or owning Stage.