Implement Parser and remove VarEnum.kt, EnumerableValue.kt
The new parser allowed defining string conversion for CommandBar.kt and FileFilter in a way that does not force them to be enumerable. This got rid of the legacy classes. Enums also now convert more leniently (case insensitive).

The parser also functions as a Constraint.kt for Config.kt and in the future will also provide fancy context-sensitive autocomplete for parser-based types.

With the parser we can enrich file filters with negation for more flexibility.