Implement ComplexCE
This editor provides smart autocompletion sequence to function as an ui builder for arbitrary (but mostly sealed) hierarchies defined by a parse definition tree populated by value/type arguments. Enables Command and FileFilter editors.

There is still work to be done to improve UX, nullability support and better API, but as is, it is fully functional, effective and with good minimalistic API