Update javafx to 19-ea+9
This brings ObservableValue.map/flatMap/orElse which overrides this project's map/flatMap, but the semantics should be exactly the same. Tests will be provided later to prove this. In the rare event that some edge case does change semantics, previous methods will be (renamed and) used to correct the behavior